+++ Flask_api/__init__.py
@@ -25,4 +25,4 @@ app.config['PERPAGE_MAX_LIMIT'] = 100
 db = SQLAlchemy(app)
 auth = HTTPBasicAuth()
 
-from Flask_api import api, models, errors, authenticate
+from Flask_api import api, models, errors, authenticate, bucketlistapi, itemapi
